+++ Radamsa/vmf/src/modules/common/mutator/RadamsaRepeatByteMutator.cpp
@@ -143,15 +143,15 @@ void RadamsaRepeatByteMutator::mutateTestCase(StorageModule& storage, StorageEnt
 
     // Select a random index from which the new bytes will be repeated.
 
-    const size_t lower{0u};
+    const unsigned long lower{0ul};
     const size_t upper{originalSize - 1u};
-    const size_t maximumRandomIndexValue{originalSize - minimumSeedIndex};
+    const unsigned long maximumRandomIndexValue{static_cast<unsigned long>(originalSize - minimumSeedIndex)};
     const size_t randomByteRepetitionIndex{
                                     std::clamp(
-                                        rand->randBetween(
+                                        static_cast<size_t>(rand->randBetween(
                                             lower,
-                                            maximumRandomIndexValue) + minimumSeedIndex,
-                                        lower,
+                                            maximumRandomIndexValue)) + minimumSeedIndex,
+                                        static_cast<size_t>(lower),
                                         upper
                                     )
     };
